refactor: 누락된 패키지명 수정
- settingsStorage -> userStorage
