+++ android/app/src/main/java/com/on/turip/data/userStorage/UserStorageMapper.kt
@@ -1,5 +1,5 @@
 package com.on.turip.data.userStorage
 
-import com.on.turip.domain.settingsStorage.TuripDeviceIdentifier
+import com.on.turip.domain.userStorage.TuripDeviceIdentifier
 
 fun String.toDomain(): TuripDeviceIdentifier = TuripDeviceIdentifier(fid = this)
